feat: se realiza ahustes en la solicitud de fecha de nacimiento

+++ src/main/java/com/bot/elara/Domain/Model/Patient.java
@@ -32,7 +32,7 @@ public class Patient {
     private String nombreCompleto; // del paciente o de la persona consultada
     private String genero; // "Femenino", "Masculino"
     private LocalDate fechaNacimiento;
-    private Integer pesoKg;
+    private Double pesoKg;
     private Double alturaM;
     private Boolean fuma;
